add CreateCheckBox component #11
