+++ src/component/styles/CreateCheckBox.module.css
@@ -1,0 +1,37 @@
+.checkbox_wrapper {
+  width: 90%;
+  margin: 0 auto;
+  display: flex;
+  flex-wrap: wrap;
+}
+
+.checkbox_text {
+  color: #000000;
+  font-size: 1.2rem;
+  border: #000000 1px solid;
+  border-radius: 5px;
+  padding: 0.3rem;
+  margin: 0.5vh 0.5vw;
+  transition: all 0.3s;
+  display: flex;
+  align-items: center;
+}
+
+.checkbox_input {
+  cursor: pointer;
+  margin: 0.3rem;
+}
+
+@media screen and (max-width: 959px) {
+  .checkbox_text {
+    font-size: 1rem;
+    margin: 0.8vh 1vw;
+  }
+}
+
+@media screen and (max-width: 480px) {
+  .checkbox_text {
+    font-size: 0.7rem;
+    margin: 1vh 1.5vw;
+  }
+}
